pyproject.toml
[project]
name = "trenkwalder-hr-chatbot"
version = "1.0.0"
description = "Trenkwalder HR Assistant — AI chatbot with tool-use and admin panel"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.9.0",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.5.0",
]
